# control-system-server/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "EntitlementOut": {
        "properties": {
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "feature_key": {
            "title": "Feature Key",
            "type": "string"
          },
          "quota": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quota"
          }
        },
        "required": [
          "feature_key",
          "enabled"
        ],
        "title": "EntitlementOut",
        "type": "object"
      },
      "EntitlementSet": {
        "properties": {
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "feature_key": {
            "title": "Feature Key",
            "type": "string"
          },
          "quota": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quota"
          }
        },
        "required": [
          "feature_key"
        ],
        "title": "EntitlementSet",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PlanOut": {
        "properties": {
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "entitlements": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/EntitlementOut"
            },
            "title": "Entitlements",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "monthly_price_minor": {
            "title": "Monthly Price Minor",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "rank": {
            "title": "Rank",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "key",
          "name",
          "description",
          "monthly_price_minor",
          "currency",
          "rank",
          "is_active"
        ],
        "title": "PlanOut",
        "type": "object"
      },
      "PlanUpdate": {
        "properties": {
          "currency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          },
          "monthly_price_minor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Monthly Price Minor"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "title": "PlanUpdate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Multi-Academy Control System",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/dashboard": {
      "get": {
        "operationId": "dashboard_v1_dashboard_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Dashboard",
        "tags": [
          "platform"
        ]
      }
    },
...
    "/v1/monitoring/audit-rate": {
      "get": {
        "operationId": "audit_rate_stub_v1_monitoring_audit_rate_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Audit Rate Stub",
        "tags": [
          "monitoring"
        ]
      }
    },
    "/v1/monitoring/load-test-notes": {
      "get": {
        "operationId": "load_test_notes_v1_monitoring_load_test_notes_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Load Test Notes",
        "tags": [
          "monitoring"
        ]
      }
    },
    "/v1/monitoring/slos": {
      "get": {
        "description": "Stub SLO / error-budget view \u2014 wire real metrics (Prometheus/Datadog) later.",
        "operationId": "slo_dashboard_v1_monitoring_slos_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Slo Dashboard",
        "tags": [
          "monitoring"
        ]
      }
    },
    "/v1/monitoring/synthetics": {
      "get": {
        "operationId": "synthetics_stub_v1_monitoring_synthetics_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Synthetics Stub",
        "tags": [
          "monitoring"
        ]
      }
    },
...
    "/v1/plans": {
      "get": {
        "operationId": "list_plans_v1_plans_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PlanOut"
                  },
                  "title": "Response List Plans V1 Plans Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Plans",
        "tags": [
          "plans"
        ]
      }
    },
    "/v1/plans/{key}": {
      "patch": {
        "operationId": "update_plan_v1_plans__key__patch",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Plan",
        "tags": [
          "plans"
        ]
      }
    },
    "/v1/plans/{key}/entitlements": {
      "put": {
        "operationId": "set_entitlement_v1_plans__key__entitlements_put",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntitlementSet"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Set Entitlement",
        "tags": [
          "plans"
        ]
      }
    },
...
  }
}